250209_v0/web_server/app/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
python-multipart
yt-dlp
SQLAlchemy
asyncpg
requests
aiohttp